Add im2im interface and rename to avoid conflicts
