+++ src/im2im/__init__.py
@@ -1,11 +1,11 @@
-from importlib.metadata import PackageNotFoundError, version  # pragma: no cover
+from importlib.metadata import PackageNotFoundError, version
 
 from .interface_py_api import _code_generator, _constructor
-from .interface_py_api import (get_conversion_code, get_convert_path, add_edge_factory_cluster,
+from .interface_py_api import (im2im, get_conversion_code, get_convert_path, add_edge_factory_cluster,
                                add_meta_values_for_image, add_conversion_for_metadata_pairs,
                                config_astar_goal_function)
 from .knowledge_graph_construction import find_closest_metadata
-from .code_exec import is_image_equal, random_test_image_and_expected, time_cost, time_cost_in_kg
+from .time_cost import is_image_equal, random_test_image_and_expected, time_cost, time_cost_in_kg
 from .metadata_differ import *
 
 try:
