+++ tests/test_py_interface.py
@@ -1,11 +1,25 @@
 import os
+
+import numpy as np
 import pytest
+import torch
 
 from src.im2im import add_conversion_for_metadata_pairs, get_convert_path, _code_generator, _constructor, \
-    get_conversion_code
+    get_conversion_code, im2im
 from src.im2im.code_generator import ConvertCodeGenerator
 from src.im2im.knowledge_graph_construction import KnowledgeGraph
 from .data_for_tests.nodes_edges import all_nodes
+
+
+def test_im2im():
+    source_image = np.random.randint(0, 256, (20, 20, 3), dtype=np.uint8)
+    expected_image = torch.from_numpy(source_image).permute(2, 0, 1).unsqueeze(0)
+
+    source_image_desc = {"lib": "numpy"}
+    target_image_desc = {"lib": "torch", "image_dtype": "uint8"}
+    actual_image = im2im(source_image, source_image_desc, target_image_desc)
+
+    assert torch.allclose(actual_image, expected_image), f"expected {expected_image} but got {actual_image}"
 
 
 @pytest.fixture
@@ -15,7 +29,7 @@ def conversion_for_metadata_pairs():
              {"color_channel": "rgb", "channel_order": "channel first", "minibatch_input": False,
               "image_data_type": "uint8", "device": "gpu", "data_representation": "torch.tensor"},
              ("", "def convert(var):\n  return var[[2, 1, 0], :, :]")),
-            ( {"color_channel": "bgr", "channel_order": "channel first", "minibatch_input": False,
+            ({"color_channel": "bgr", "channel_order": "channel first", "minibatch_input": False,
               "image_data_type": "uint8", "device": "gpu", "data_representation": "torch.tensor"},
              {"color_channel": "rgb", "channel_order": "channel first", "minibatch_input": False,
               "image_data_type": "uint8", "device": "gpu", "data_representation": "torch.tensor"},
@@ -26,6 +40,7 @@ def conversion_for_metadata_pairs():
 def test_add_conversion_for_metadata_pair_single_value(conversion_for_metadata_pairs):
     def noop():
         pass
+
     _constructor.save_knowledge_graph = noop
 
     _constructor.clear_knowledge_graph()
@@ -41,6 +56,7 @@ def test_add_conversion_for_metadata_pair_single_value(conversion_for_metadata_p
 def test_add_conversion_for_metadata_pair_list_values(conversion_for_metadata_pairs):
     def noop():
         pass
+
     _constructor.save_knowledge_graph = noop
 
     _constructor.clear_knowledge_graph()
@@ -57,6 +73,7 @@ def test_add_conversion_for_metadata_pair_list_values(conversion_for_metadata_pa
 def test_add_conversion_for_metadata_pair_empty():
     def noop():
         pass
+
     _constructor.save_knowledge_graph = noop
 
     _constructor.clear_knowledge_graph()
@@ -68,6 +85,7 @@ def test_add_conversion_for_metadata_pair_empty():
 def test_add_conversion_for_metadata_pair_none():
     def noop():
         pass
+
     _constructor.save_knowledge_graph = noop
 
     _constructor.clear_knowledge_graph()
